+++ example/src/main/scala/example/HelloWorldWithMiddlewares.scala
@@ -1,7 +1,8 @@
 package example
 
-import zhttp.http.HttpMiddleware.{addHeader, debug, patchZIO, timeout}
 import zhttp.http._
+import zhttp.http.middleware.HttpMiddleware
+import zhttp.http.middleware.Middleware._
 import zhttp.service.Server
 import zio.clock.{Clock, currentTime}
 import zio.console.Console
