+++ zio-http/src/main/scala/zhttp/http/package.scala
@@ -6,14 +6,13 @@ import zio.ZIO
 import java.nio.charset.Charset
 
 package object http extends PathModule with RequestSyntax with RouteDecoderModule {
-  type HttpApp[-R, +E]        = Http[R, E, Request, Response]
-  type UHttpApp               = HttpApp[Any, Nothing]
-  type RHttpApp[-R]           = HttpApp[R, Throwable]
-  type UHttp[-A, +B]          = Http[Any, Nothing, A, B]
-  type SilentResponse[-E]     = CanBeSilenced[E, Response]
-  type ResponseZIO[-R, +E]    = ZIO[R, E, Response]
-  type Header                 = (CharSequence, CharSequence)
-  type HttpMiddleware[-R, +E] = Middleware[R, E, Request, Response, Request, Response]
+  type HttpApp[-R, +E]     = Http[R, E, Request, Response]
+  type UHttpApp            = HttpApp[Any, Nothing]
+  type RHttpApp[-R]        = HttpApp[R, Throwable]
+  type UHttp[-A, +B]       = Http[Any, Nothing, A, B]
+  type SilentResponse[-E]  = CanBeSilenced[E, Response]
+  type ResponseZIO[-R, +E] = ZIO[R, E, Response]
+  type Header              = (CharSequence, CharSequence)
 
   /**
    * Default HTTP Charset
