+++ zio-http/src/test/scala/zhttp/http/middleware/MiddlewareExtensionsSpec.scala
@@ -1,26 +1,26 @@
-package zhttp.middleware
+package zhttp.http.middleware
 
 import zhttp.http._
+import zhttp.http.middleware.Middleware._
 import zhttp.internal.HttpAppTestExtensions
 import zio._
 import zio.clock.Clock
 import zio.duration._
 import zio.test.Assertion._
+import zio.test._
 import zio.test.environment.{TestClock, TestConsole}
-import zio.test.{DefaultRunnableSpec, assert, assertM}
 
-object HttpMiddlewareSpec extends DefaultRunnableSpec with HttpAppTestExtensions {
+object MiddlewareExtensionsSpec extends DefaultRunnableSpec with HttpAppTestExtensions {
   private val app: HttpApp[Any with Clock, Nothing] = Http.collectZIO[Request] { case Method.GET -> !! / "health" =>
     UIO(Response.ok).delay(1 second)
   }
-  private val midA                                  = HttpMiddleware.addHeader("X-Custom", "A")
-  private val midB                                  = HttpMiddleware.addHeader("X-Custom", "B")
+  private val midA                                  = Middleware.addHeader("X-Custom", "A")
+  private val midB                                  = Middleware.addHeader("X-Custom", "B")
   private val basicHS                               = Headers.basicAuthorizationHeader("user", "resu")
   private val basicHF                               = Headers.basicAuthorizationHeader("user", "user")
-  private val basicAuthM = HttpMiddleware.basicAuth { case (u, p) => p.toString.reverse == u }
+  private val basicAuthM                            = Middleware.basicAuth { case (u, p) => p.toString.reverse == u }
 
   def spec = suite("HttpMiddleware") {
-    import HttpMiddleware._
 
     suite("debug") {
       testM("log status method url and time") {
@@ -69,7 +69,7 @@ object HttpMiddlewareSpec extends DefaultRunnableSpec with HttpAppTestExtensions
             assertM(program(Request()))(isNone)
           }
       } +
-      suite("ifThenElseM") {
+      suite("ifThenElseZIO") {
         testM("if the condition is true take first") {
           val app = (Http.ok @@ ifThenElseZIO(condM(true))(midA, midB)) getHeader "X-Custom"
           assertM(app(Request()))(isSome(equalTo("A")))
@@ -102,11 +102,11 @@ object HttpMiddlewareSpec extends DefaultRunnableSpec with HttpAppTestExtensions
       } +
       suite("when") {
         testM("if the condition is true apple middleware") {
-          val app = Http.ok @@ when(cond(true))(midA) getHeader "X-Custom"
+          val app = Http.ok @@ Middleware.when(cond(true))(midA) getHeader "X-Custom"
           assertM(app(Request()))(isSome(equalTo("A")))
         } +
           testM("if the condition is false don't apply the middleware") {
-            val app = Http.ok @@ when(cond(false))(midA) getHeader "X-Custom"
+            val app = Http.ok @@ Middleware.when(cond(false))(midA) getHeader "X-Custom"
             assertM(app(Request()))(isNone)
           }
       } +
@@ -126,50 +126,6 @@ object HttpMiddlewareSpec extends DefaultRunnableSpec with HttpAppTestExtensions
             }
         }
       } +
-      suite("cors") {
-        // FIXME:The test should ideally pass with `Http.ok` also
-        val app = Http.collect[Request] { case Method.GET -> !! / "success" => Response.ok } @@ cors()
-        testM("OPTIONS request") {
-          val request = Request(
-            method = Method.OPTIONS,
-            url = URL(!! / "success"),
-            headers = Headers.accessControlRequestMethod(Method.GET) ++ Headers.origin("test-env"),
-          )
-
-          val expected = Headers
-            .accessControlAllowCredentials(true)
-            .withAccessControlAllowMethods(Method.GET)
-            .withAccessControlAllowOrigin("test-env")
-            .withAccessControlAllowHeaders(
-              CORS.DefaultCORSConfig.allowedHeaders.getOrElse(Set.empty).mkString(","),
-            )
-            .toList
-
-          for {
-            res <- app(request)
-          } yield assert(res.getHeadersAsList)(hasSubset(expected)) &&
-            assert(res.status)(equalTo(Status.NO_CONTENT))
-        } +
-          testM("GET request") {
-            val request =
-              Request(
-                method = Method.GET,
-                url = URL(!! / "success"),
-                headers = Headers.accessControlRequestMethod(Method.GET) ++ Headers.origin("test-env"),
-              )
-
-            val expected = Headers
-              .accessControlExposeHeaders("*")
-              .withAccessControlAllowOrigin("test-env")
-              .withAccessControlAllowMethods(Method.GET)
-              .withAccessControlAllowCredentials(true)
-              .toList
-
-            for {
-              res <- app(request)
-            } yield assert(res.getHeadersAsList)(hasSubset(expected))
-          }
-      } +
       suite("cookie") {
         testM("addCookie") {
           val cookie = Cookie("test", "testValue")
@@ -185,33 +141,6 @@ object HttpMiddlewareSpec extends DefaultRunnableSpec with HttpAppTestExtensions
             assertM(app(Request()))(
               equalTo(Some(cookie.encode)),
             )
-          }
-      } +
-      suite("csrf") {
-        val app           = (Http.ok @@ csrfValidate("x-token")).getStatus
-        val setCookie     = Headers.cookie(Cookie("x-token", "secret"))
-        val invalidXToken = Headers("x-token", "secret1")
-        val validXToken   = Headers("x-token", "secret")
-        testM("x-token not present") {
-          assertM(app(Request(headers = setCookie)))(equalTo(Status.FORBIDDEN))
-        } +
-          testM("x-token mismatch") {
-            assertM(app(Request(headers = setCookie ++ invalidXToken)))(
-              equalTo(Status.FORBIDDEN),
-            )
-          } +
-          testM("x-token match") {
-            assertM(app(Request(headers = setCookie ++ validXToken)))(
-              equalTo(Status.OK),
-            )
-          } +
-          testM("app execution skipped") {
-            for {
-              r <- Ref.make(false)
-              app = Http.ok.tapZIO(_ => r.set(true)) @@ csrfValidate("x-token")
-              _   <- app(Request(headers = setCookie ++ invalidXToken))
-              res <- r.get
-            } yield assert(res)(equalTo(false))
           }
       }
   }
